初步完成了/room/match接口
进行了很少的测试，因为match需要多人登录，模拟器来比较麻烦，就暂时没测试。
更改了一下RoomCreateController，RoomID统一用int。
另外，移动了一下配置项的位置，把自定义的配置都写在了application.yml中，如果对于某个profile，有哪个参数需要调整，就在对应的配置文件（比如application-develop.yml）中直接重写（覆盖）掉就行了。

+++ src/main/java/cn/sduonline/sdu_lantern_festival_2021/restcontroller/room/RoomCreateController.java
@@ -28,7 +28,7 @@ public class RoomCreateController {
             return Response.fail(ResponseCode.ERROR_PARAMETER_INVALID_VALUE);
 
         // 创建一个房间（房间id自增）
-        long roomID = redisUtil.incr("room_id_incr", 1);
+        int roomID = (int) redisUtil.incr("room_id_incr", 1);
 
         Map<String, Object> attributes = new HashMap<>();
         attributes.put("room_id", roomID);
